Create code
P1. Design method to read a record from file.
P2. Verify card number is valid. Create an instance for class.
P3. I think factory method or adapter patterns to be most suited for this problem.
P4. A consequence of using either of this patterns might be that thy are too big or complicated/unecessary for a small array of strings.